requirements.txt
opencv-python
ultralytics
deepface
psycopg2-binary
fastapi
uvicorn
python-multipart